requirements.txt
numpy
Pillow
scipy
tqdm
graspnetAPI
open3d
torch
torchvision